requirements.txt
transformers>=4.49.0
wandb>=0.17.0
accelerate
torch
pillow
numpy
easyocr
paddleocr
pytesseract
opencv-python
tqdm
pandas
datasets
# Optional extras:
# paddleocr
# timm
